added board revision versioning
Hardware is being added to the v2 board:

* A chip for accurate spindle speed control
* An EEPROM chip for eventually storing version/calibration data

To address these hardware additions, changes had to be made to the
hal and ini settings for MachineKit. Since we needed to also support
the previous board revision of the v2, we restructured the settings
a bit to allow for a specific board revision to pull specific settings.
This allows us to also support the v1.