v4.0.0
Feature overlays:
Support for the high speed spindle and other features coming down the line... stay tuned!

Improved G code program upload/download:
Prior versions of the UI could only handle files about 2.5MB in size. With this update the UI can handle files over 100MB in size with upload and download indicators to show progress. This allows the Pocket NC to handle programs with millions of lines of G code.

Spindle clocking:
For machines with the necessary hardware (machines released around June 2018), the measured spindle speed will be shown in the DRO panel. If the machine doesn't have the necessary hardware (the V1 and V2's released prior to June 2018), the DRO will show the commanded speed with any spindle speed overrides applied.

Improved metric jogging increments:
The Jog widget now has sensical fixed step sizes when in metric mode (1mm, .1mm, .01mm and .001mm).

Improved MDI interface:
An indicator was added to the MDI input to show when an MDI command was actually sent to the Pocket NC (a green check indicates the command was successfully sent to the Pocket NC). Better error notifications are shown when MDI commands are not successful or are not able to be sent.

Improved Work Offsets panel:
Fixed bugs with the Work Offsets panel that would trigger a refresh of the UI and/or to improperly set work offsets.

Improved workflow for connecting over ethernet:
The Pocket NC can be connected to a network using its ethernet port, which is required to receive updates. You can also connect to the Pocket NC UI over the network, but it required changing settings in the Conf tab for it to work. Now you can connect to the Pocket NC UI without needing that extra step.